# Conflicts:
#	flinkx-connectors/flinkx-connector-jdbc-base/src/main/java/com/dtstack/flinkx/connector/jdbc/sink/JdbcOutputFormat.java
#	flinkx-connectors/flinkx-connector-phoenix5/src/main/java/com/dtstack/flinkx/connector/phoenix5/sink/Phoenix5OutputFormat.java
#	flinkx-core/src/main/java/com/dtstack/flinkx/sink/format/BaseRichOutputFormat.java

+++ flinkx-connectors/flinkx-connector-oraclelogminer/src/main/java/com/dtstack/flinkx/connector/oraclelogminer/listener/PositionManager.java
@@ -16,8 +16,9 @@
  * limitations under the License.
  */
 
-
 package com.dtstack.flinkx.connector.oraclelogminer.listener;
+
+import java.math.BigInteger;
 
 /**
  * @author jiangbo
@@ -25,13 +26,13 @@ package com.dtstack.flinkx.connector.oraclelogminer.listener;
  */
 public class PositionManager {
 
-    private Long scn;
+    private BigInteger scn;
 
-    public void updatePosition(Long scn) {
+    public void updatePosition(BigInteger scn) {
         this.scn = scn;
     }
 
-    public Long getPosition() {
+    public BigInteger getPosition() {
         return this.scn;
     }
 }
